moved Header component into separate file
+removed unused styling, commented out for now in case it's needed, eg in mobile view?
