+++ packages/components/transcript-editor/index.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
-  faCog,
   faKeyboard,
   faQuestionCircle,
   faMousePointer,
@@ -10,6 +8,7 @@ import {
   faUserEdit,
   faSave
 } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
 import Tooltip from 'react-simple-tooltip';
 
@@ -19,47 +18,9 @@ import VideoPlayer from '../video-player';
 import Settings from '../settings';
 import Shortcuts from '../keyboard-shortcuts';
 import { secondsToTimecode } from '../../util/timecode-converter';
+import Header from './src/Header.js';
 
 import style from './index.module.css';
-
-// const Header = (props) => {
-class Header extends React.Component {
-  shouldComponentUpdate(nextProps) {
-    return false;
-  }
-  render() {
-    const props = this.props;
-
-    return ( <>
-      <header className={ style.header }>
-        {props.showSettings ? props.settings : null}
-        {props.showShortcuts ? props.shortcuts : null}
-        {props.tooltip}
-      </header>
-      <nav className={ style.nav }>
-        {props.mediaUrl === null ? null : props.mediaControls}
-      </nav>
-      <div className={ style.settingsContainer }>
-        <button
-          className={ style.settingsButton }
-          title="Settings"
-          onClick={ props.handleSettingsToggle }
-        >
-          <FontAwesomeIcon icon={ faCog } />
-        </button>
-        <button
-          className={ `${ style.settingsButton } ${
-            style.keyboardShortcutsButon
-          }` }
-          title="view shortcuts"
-          onClick={ props.handleShortcutsToggle }
-        >
-          <FontAwesomeIcon icon={ faKeyboard } />
-        </button>
-      </div>
-    </>);
-  };
-}
 
 class TranscriptEditor extends React.Component {
   constructor(props) {
