+++ packages/components/transcript-editor/src/Header.js
@@ -1,0 +1,56 @@
+import React from 'react';
+import {
+  faCog,
+  faKeyboard
+} from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+import style from '../index.module.css';
+
+class Header extends React.Component {
+
+  shouldComponentUpdate(nextProps) {
+    if (nextProps.showSettings !== this.props.showSettings) {
+      return true;
+    }
+    if (nextProps.showShortcuts !== this.props.showShortcuts) {
+      return true;
+    }
+
+    return false;
+  }
+  render() {
+    const props = this.props;
+
+    return (<>
+      <header className={ style.header }>
+        {props.showSettings ? props.settings : null}
+        {props.showShortcuts ? props.shortcuts : null}
+        {props.tooltip}
+      </header>
+      <nav className={ style.nav }>
+        {props.mediaUrl === null ? null : props.mediaControls}
+      </nav>
+      <div className={ style.settingsContainer }>
+        <button
+          className={ style.settingsButton }
+          title="Settings"
+          onClick={ props.handleSettingsToggle }
+        >
+          <FontAwesomeIcon icon={ faCog } />
+        </button>
+        <button
+          className={ `${ style.settingsButton } ${
+            style.keyboardShortcutsButon
+          }` }
+          title="view shortcuts"
+          onClick={ props.handleShortcutsToggle }
+        >
+          <FontAwesomeIcon icon={ faKeyboard } />
+        </button>
+      </div>
+    </>);
+  };
+}
+
+export default Header;
